requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
pytorch-lightning>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0

# Transformers for Lyrics
transformers>=4.30.0
sentencepiece>=0.1.99
tokenizers>=0.13.0
langdetect>=1.0.9
translators>=5.8.0

# Clustering and ML
scikit-learn>=1.3.0
scipy>=1.11.0
umap-learn>=0.5.3
hdbscan>=0.8.33

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
tensorboard>=2.13.0

# Evaluation
scikit-learn-extra>=0.3.0

# Optimization
accelerate>=0.20.0
bitsandbytes>=0.41.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
wandb>=0.15.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# File Processing
pathlib>=1.0.1
joblib>=1.3.0

# Audio Augmentation
audiomentations>=0.31.0

# Multi-language Support
pycountry>=22.3.5
polyglot>=16.7.4
